requirements.txt
Django==5.2.8
celery==5.3.4
redis==5.0.1
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0